+++ src/test/java/com/braintreegateway/integrationtest/SepaDirectDebitAccountIT.java
@@ -73,6 +73,40 @@ public class SepaDirectDebitAccountIT extends IntegrationTest {
     }
 
     @Test
+    public void findReturnsSubscriptionsAssociatedWithASepaDirectDebitAccount() {
+        String nonce = TestHelper.generateSepaDebitNonce(gateway);
+        Result<Customer> customerResult = gateway.customer().create(new CustomerRequest());
+        assertTrue(customerResult.isSuccess());
+        Customer customer = customerResult.getTarget();
+
+        PaymentMethodRequest request = new PaymentMethodRequest().
+            customerId(customer.getId()).
+            paymentMethodNonce(nonce);
+
+        Result<? extends PaymentMethod> result = gateway.paymentMethod().create(request);
+        assertTrue(result.isSuccess());
+
+        PaymentMethod paymentMethod = result.getTarget();
+
+        Plan plan = PlanFixture.PLAN_WITHOUT_TRIAL;
+        SubscriptionRequest subscriptionRequest = new SubscriptionRequest().
+                paymentMethodToken(paymentMethod.getToken()).
+                planId(plan.getId());
+
+        Result<Subscription> subscriptionResult = gateway.subscription().create(subscriptionRequest);
+        assertTrue(subscriptionResult.isSuccess());
+        Subscription subscription1 = subscriptionResult.getTarget();
+        subscriptionResult = gateway.subscription().create(subscriptionRequest);
+        assertTrue(subscriptionResult.isSuccess());
+        Subscription subscription2 = subscriptionResult.getTarget();
+
+        SepaDirectDebitAccount found = gateway.sepaDirectDebitAccount().find(paymentMethod.getToken());
+        assertNotNull(found);
+        assertTrue(found.getSubscriptions().contains(subscription1));
+        assertTrue(found.getSubscriptions().contains(subscription2));
+    }
+
+    @Test
     public void findThrowsNotFoundExceptionWhenSepaDebitAccountIsMissing() {
         try {
             gateway.sepaDirectDebitAccount().find(" ");